+++ wavespectra/construct/__init__.py
@@ -145,7 +145,7 @@ def partition_and_reconstruct(
             if v not in dset.data_vars:
                 raise ValueError(f"Missing variable '{v}' in dset for partitioning.")
         kw = dict(wspd=dset.wspd, wdir=dset.wdir, dpt=dset.dpt, swells=parts - 1)
-    dspart = getattr(dset.spec.partition, partition_method)(**kw)
+    dspart = getattr(dset[attrs.SPECNAME].spec.partition, partition_method)(**kw)
 
     # Calculating parameters
     stats = list(set(STATS) - set(use_defaults))